librustc: Move inline asm stuff to different mod.
